renomer la classe SortieDAOImplementation.java en JpaSortieDao.java et ajouter une méthode

+++ src/main/java/dao/CategorieDAOImplementation.java
@@ -22,7 +22,7 @@ public class CategorieDAOImplementation implements CategorieDAO {
     }
 
     @Override
-    public Categorie findById(Integer id) {
+    public Categorie findById(long id) {
         return categorieRepository.findById(id).orElse(null);
     }
